refactor: extract TradeDialog and Sparkline into reusable components to eliminate code duplication, move profile card above ratings, and match headings styling
