+++ react_main/src/components/TradeDialog.jsx
@@ -1,0 +1,284 @@
+import React, { useState, useMemo, useContext } from "react";
+import axios from "axios";
+import {
+  Dialog,
+  DialogTitle,
+  DialogContent,
+  DialogActions,
+  Stack,
+  Box,
+  Typography,
+  ToggleButton,
+  ToggleButtonGroup,
+  Button,
+  TextField,
+  Paper
+} from "@mui/material";
+import { UserContext, SiteInfoContext } from "Contexts";
+import { useErrorAlert } from "components/Alerts";
+import { Avatar } from "pages/User/User"; // Reuse existing custom Avatar component
+import { Icon } from "@iconify/react";
+
+export default function TradeDialog({ open, onClose, stock, initialType = "buy", onSuccess }) {
+  const user = useContext(UserContext);
+  const siteInfo = useContext(SiteInfoContext);
+  const errorAlert = useErrorAlert();
+
+  const [tradeType, setTradeType] = useState(initialType);
+  const [shareCount, setShareCount] = useState(1);
+  const [submitting, setSubmitting] = useState(false);
+
+  // Reset modal state when opening with a new stock
+  React.useEffect(() => {
+    if (open) {
+      setTradeType(initialType || "buy");
+      setShareCount(1);
+    }
+  }, [open, stock, initialType]);
+
+  // Calculate pricing preview for modal on bonding curve
+  const tradePreview = useMemo(() => {
+    if (!stock || shareCount <= 0) {
+      return { price: 0, creatorFee: 0, systemFee: 0, total: 0 };
+    }
+
+    const currentSupply = stock.shareSupply;
+    let basePrice = 0;
+
+    if (tradeType === "buy") {
+      for (let i = 1; i <= shareCount; i++) {
+        const S = currentSupply + i;
+        basePrice += Math.max(1, Math.floor((S * S) / 100));
+      }
+      const creatorFee = Math.max(1, Math.round(basePrice * 0.05));
+      const systemFee = Math.max(1, Math.round(basePrice * 0.05));
+      return {
+        price: basePrice,
+        creatorFee,
+        systemFee,
+        total: basePrice + creatorFee + systemFee,
+      };
+    } else {
+      const sellCount = Math.min(shareCount, currentSupply);
+      for (let i = 0; i < sellCount; i++) {
+        const S = currentSupply - i;
+        basePrice += Math.max(1, Math.floor((S * S) / 100));
+      }
+      const creatorFee = Math.max(1, Math.round(basePrice * 0.05));
+      const systemFee = Math.max(1, Math.round(basePrice * 0.05));
+      return {
+        price: basePrice,
+        creatorFee,
+        systemFee,
+        total: Math.max(0, basePrice - creatorFee - systemFee),
+      };
+    }
+  }, [stock, tradeType, shareCount]);
+
+  if (!stock) return null;
+
+  const handleConfirmTrade = () => {
+    if (shareCount <= 0) return;
+    setSubmitting(true);
+
+    const isPlayer = stock.targetType === "player";
+    const endpoint = isPlayer
+      ? (tradeType === "buy" ? "/api/stocks/buy" : "/api/stocks/sell")
+      : (tradeType === "buy" ? "/api/stocks/families/buy" : "/api/stocks/families/sell");
+
+    const payload = isPlayer
+      ? { subjectId: stock.id, shares: shareCount }
+      : { familyId: stock.id, shares: shareCount };
+
+    axios
+      .post(endpoint, payload)
+      .then((res) => {
+        const msg =
+          tradeType === "buy"
+            ? `Successfully bought ${shareCount} shares for ${res.data.total} coins.`
+            : `Successfully sold ${shareCount} shares for ${res.data.totalPayout} coins.`;
+        siteInfo.showAlert(msg, "success");
+        onClose();
+        if (onSuccess) onSuccess();
+      })
+      .catch(errorAlert)
+      .finally(() => setSubmitting(false));
+  };
+
+  const isBuyDisabled = stock.targetType === "player" && user.id === stock.id;
+
+  return (
+    <Dialog open={open} onClose={onClose} fullWidth maxWidth="xs">
+      <DialogTitle sx={{ fontWeight: "bold", borderBottom: 1, borderColor: "divider", pb: 2 }}>
+        <Stack direction="row" spacing={1.5} alignItems="center">
+          {stock.targetType === "player" ? (
+            <Avatar id={stock.id} name={stock.name} avatar={stock.avatar} size={40} />
+          ) : (
+            <Box
+              sx={{
+                width: 40,
+                height: 40,
+                borderRadius: "50%",
+                backgroundImage: stock.avatar ? `url(/uploads/${stock.id}_family_avatar.webp)` : "none",
+                backgroundSize: "cover",
+                backgroundPosition: "center",
+                display: "flex",
+                alignItems: "center",
+                justifyContent: "center",
+                background: stock.avatar ? "none" : "rgba(255, 255, 255, 0.05)",
+                border: "1px solid rgba(255, 255, 255, 0.12)",
+              }}
+            >
+              {!stock.avatar && <Icon icon="lucide:users" style={{ fontSize: "20px" }} />}
+            </Box>
+          )}
+          <Box>
+            <Typography variant="h6" sx={{ lineHeight: 1.2 }}>
+              {stock.name}
+            </Typography>
+            <Typography variant="caption" color="text.secondary">
+              Current share supply: {stock.shareSupply}
+            </Typography>
+          </Box>
+        </Stack>
+      </DialogTitle>
+
+      <DialogContent sx={{ py: 3 }}>
+        <Stack spacing={3}>
+          <ToggleButtonGroup
+            value={tradeType}
+            exclusive
+            onChange={(e, val) => {
+              if (val) {
+                setTradeType(val);
+                setShareCount(1);
+              }
+            }}
+            fullWidth
+            color={tradeType === "buy" ? "success" : "error"}
+          >
+            <ToggleButton value="buy" disabled={isBuyDisabled}>
+              Buy Shares
+            </ToggleButton>
+            <ToggleButton value="sell">Sell Shares</ToggleButton>
+          </ToggleButtonGroup>
+
+          <Box>
+            <Typography variant="subtitle2" sx={{ mb: 1 }}>
+              Number of Shares
+            </Typography>
+            <Stack direction="row" spacing={1} alignItems="center">
+              <Button
+                variant="outlined"
+                onClick={() => setShareCount((c) => Math.max(1, c - 1))}
+                sx={{ minWidth: 48 }}
+              >
+                -
+              </Button>
+              <TextField
+                fullWidth
+                type="number"
+                value={shareCount}
+                onChange={(e) => {
+                  const num = Number(e.target.value);
+                  if (num > 0) setShareCount(Math.floor(num));
+                }}
+                slotProps={{ htmlInput: { min: 1, style: { textAlign: "center" } } }}
+              />
+              <Button
+                variant="outlined"
+                onClick={() => setShareCount((c) => c + 1)}
+                sx={{ minWidth: 48 }}
+              >
+                +
+              </Button>
+            </Stack>
+            {tradeType === "sell" && (
+              <Typography variant="caption" color="text.secondary" display="block" sx={{ mt: 0.5 }}>
+                You own {stock.sharesOwned} shares.
+              </Typography>
+            )}
+          </Box>
+
+          <Paper variant="outlined" sx={{ p: 2, background: "rgba(255,255,255,0.01)" }}>
+            <Stack spacing={1}>
+              <Stack direction="row" justifyContent="space-between">
+                <Typography variant="body2">Base Price:</Typography>
+                <Typography variant="body2" sx={{ fontWeight: "bold" }}>
+                  {tradePreview.price} Coins
+                </Typography>
+              </Stack>
+              <Stack direction="row" justifyContent="space-between">
+                <Typography variant="body2">
+                  {stock.targetType === "player" ? "Creator Fee (5%):" : "Treasury Fee (5%):"}
+                </Typography>
+                <Typography variant="body2" color="success.main">
+                  +{tradePreview.creatorFee} Coins
+                </Typography>
+              </Stack>
+              <Stack direction="row" justifyContent="space-between">
+                <Typography variant="body2">System Fee (5%):</Typography>
+                <Typography variant="body2" color="warning.main">
+                  +{tradePreview.systemFee} Coins
+                </Typography>
+              </Stack>
+              <Box sx={{ borderTop: 1, borderColor: "divider", my: 1, pt: 1 }} />
+              <Stack direction="row" justifyContent="space-between">
+                <Typography sx={{ fontWeight: "bold" }}>
+                  {tradeType === "buy" ? "Total Cost:" : "Total Payout:"}
+                </Typography>
+                <Typography variant="h6" sx={{ fontWeight: "bold", color: "gold" }}>
+                  {tradePreview.total} Coins
+                </Typography>
+              </Stack>
+            </Stack>
+          </Paper>
+
+          {user.loggedIn && (
+            <Stack direction="row" justifyContent="space-between" sx={{ px: 1 }}>
+              <Typography variant="caption" color="text.secondary">
+                Coins After Trade:
+              </Typography>
+              <Typography
+                variant="caption"
+                sx={{
+                  fontWeight: "bold",
+                  color:
+                    tradeType === "buy"
+                      ? (user.coins || 0) >= tradePreview.total
+                        ? "success.main"
+                        : "error.main"
+                      : "success.main",
+                }}
+              >
+                {tradeType === "buy"
+                  ? ((user.coins || 0) - tradePreview.total).toFixed(2)
+                  : ((user.coins || 0) + tradePreview.total).toFixed(2)}{" "}
+                Coins
+              </Typography>
+            </Stack>
+          )}
+        </Stack>
+      </DialogContent>
+
+      <DialogActions sx={{ borderTop: 1, borderColor: "divider", p: 2 }}>
+        <Button onClick={onClose} disabled={submitting}>
+          Cancel
+        </Button>
+        <Button
+          variant="contained"
+          color={tradeType === "buy" ? "success" : "error"}
+          onClick={handleConfirmTrade}
+          disabled={
+            submitting ||
+            shareCount <= 0 ||
+            (tradeType === "buy" && (user.coins || 0) < tradePreview.total) ||
+            (tradeType === "sell" && stock.sharesOwned < shareCount)
+          }
+        >
+          {submitting ? "Trading..." : tradeType === "buy" ? "Buy" : "Sell"}
+        </Button>
+      </DialogActions>
+    </Dialog>
+  );
+}
