+++ react_main/src/pages/User/Profile.jsx
@@ -40,6 +40,8 @@ import RapSheet from "../../components/RapSheet";
 import TrophyCase from "components/TrophyCase";
 import { AchievementPanel } from "components/Achievements";
 import Scrapbook from "components/Scrapbook";
+import TradeDialog from "components/TradeDialog";
+import Sparkline from "components/Sparkline";
 import PendingTradeConfirmations from "components/PendingTradeConfirmations";
 import CasePanel from "components/CasePanel";
 import { RoleCount } from "components/Roles";
@@ -77,51 +79,7 @@ export const PRESTIGE_ICON = require(`images/prestige.png`);
 export const ACHIEVEMENTS_ICON = require(`images/achievements.png`);
 export const DAILY_ICON = require(`images/dailyChallenges.png`);
 
-// Render a sleek sparkline trend curve based on historical price ticks
-function Sparkline({ history }) {
-  const points = React.useMemo(() => {
-    let prices = Array.isArray(history) ? history : [];
-    if (prices.length === 0) prices = [1, 1];
-    if (prices.length === 1) prices = [prices[0], prices[0]];
 
-    const min = Math.min(...prices);
-    const max = Math.max(...prices);
-    const range = max - min === 0 ? 1 : max - min;
-
-    const width = 140;
-    const height = 40;
-    const padding = 4;
-
-    return prices.map((val, idx) => {
-      const x = (idx / (prices.length - 1)) * (width - padding * 2) + padding;
-      const y = height - (((val - min) / range) * (height - padding * 2) + padding);
-      return `${x},${y}`;
-    }).join(" ");
-  }, [history]);
-
-  const isUp = React.useMemo(() => {
-    let prices = Array.isArray(history) ? history : [];
-    if (prices.length < 2) return true;
-    return prices[prices.length - 1] >= prices[prices.length - 2];
-  }, [history]);
-
-  const strokeColor = isUp ? "#4caf50" : "#f44336"; // Green vs red
-
-  return (
-    <Box sx={{ display: "inline-flex", alignItems: "center" }}>
-      <svg width="140" height="40" style={{ overflow: "visible" }}>
-        <polyline
-          fill="none"
-          stroke={strokeColor}
-          strokeWidth="2.5"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          points={points}
-        />
-      </svg>
-    </Box>
-  );
-}
 
 function FavoritedRolesPanel({
   favoriteRoles = [],
@@ -267,9 +225,6 @@ export default function Profile() {
   // Stock Trading States
   const [stockInfo, setStockInfo] = useState(null);
   const [tradeModalOpen, setTradeModalOpen] = useState(false);
-  const [tradeType, setTradeType] = useState("buy");
-  const [shareCount, setShareCount] = useState(1);
-  const [tradeSubmitting, setTradeSubmitting] = useState(false);
 
   const user = useContext(UserContext);
   const siteInfo = useContext(SiteInfoContext);
@@ -283,44 +238,7 @@ export default function Profile() {
   const isBlocked = !isSelf && user.blockedUsers.indexOf(profileUserId) !== -1;
   const canViewNameHistory = user.perms.seeModPanel;
 
-  // Calculate pricing preview for modal (called unconditionally before redirects/early returns)
-  const tradePreview = useMemo(() => {
-    if (!stockInfo || shareCount <= 0) {
-      return { price: 0, creatorFee: 0, systemFee: 0, total: 0 };
-    }
 
-    const currentSupply = stockInfo.shareSupply;
-    let basePrice = 0;
-
-    if (tradeType === "buy") {
-      for (let i = 1; i <= shareCount; i++) {
-        const S = currentSupply + i;
-        basePrice += Math.max(1, Math.floor((S * S) / 100));
-      }
-      const creatorFee = Math.max(1, Math.round(basePrice * 0.05));
-      const systemFee = Math.max(1, Math.round(basePrice * 0.05));
-      return {
-        price: basePrice,
-        creatorFee,
-        systemFee,
-        total: basePrice + creatorFee + systemFee,
-      };
-    } else {
-      const sellCount = Math.min(shareCount, currentSupply);
-      for (let i = 0; i < sellCount; i++) {
-        const S = currentSupply - i;
-        basePrice += Math.max(1, Math.floor((S * S) / 100));
-      }
-      const creatorFee = Math.max(1, Math.round(basePrice * 0.05));
-      const systemFee = Math.max(1, Math.round(basePrice * 0.05));
-      return {
-        price: basePrice,
-        creatorFee,
-        systemFee,
-        total: Math.max(0, basePrice - creatorFee - systemFee),
-      };
-    }
-  }, [stockInfo, tradeType, shareCount]);
 
   // userId is the id of the current profile
   // user.id is the id of the current user
@@ -1168,186 +1086,8 @@ export default function Profile() {
   if (!profileLoaded || !user.loaded) return <Loading small />;
 
   const handleOpenTrade = () => {
-    setTradeType("buy");
-    setShareCount(1);
     setTradeModalOpen(true);
   };
-
-  const handleConfirmTrade = () => {
-    if (shareCount <= 0 || !stockInfo) return;
-    setTradeSubmitting(true);
-
-    const endpoint = tradeType === "buy" ? "/api/stocks/buy" : "/api/stocks/sell";
-    axios
-      .post(endpoint, { subjectId: profileUserId, shares: shareCount })
-      .then((res) => {
-        const msg =
-          tradeType === "buy"
-            ? `Successfully bought ${shareCount} shares for ${res.data.total} coins.`
-            : `Successfully sold ${shareCount} shares for ${res.data.totalPayout} coins.`;
-        siteInfo.showAlert(msg, "success");
-        setTradeModalOpen(false);
-        setProfileRefetchKey((k) => k + 1);
-      })
-      .catch((err) => {
-        setTradeSubmitting(false);
-        errorAlert(err);
-      })
-      .finally(() => setTradeSubmitting(false));
-  };
-
-  const tradeModal = stockInfo && (
-    <Dialog open={tradeModalOpen} onClose={() => setTradeModalOpen(false)} fullWidth maxWidth="xs">
-      <DialogTitle sx={{ fontWeight: "bold", borderBottom: 1, borderColor: "divider", pb: 2 }}>
-        <Stack direction="row" spacing={1.5} alignItems="center">
-          <Avatar id={profileUserId} name={name} avatar={avatar} size={40} />
-          <Box>
-            <Typography variant="h6" sx={{ lineHeight: 1.2 }}>
-              {name}'s Stock
-            </Typography>
-            <Typography variant="caption" color="text.secondary">
-              Current share supply: {stockInfo.shareSupply}
-            </Typography>
-          </Box>
-        </Stack>
-      </DialogTitle>
-
-      <DialogContent sx={{ py: 3 }}>
-        <Stack spacing={3}>
-          <ToggleButtonGroup
-            value={tradeType}
-            exclusive
-            onChange={(e, val) => {
-              if (val) {
-                setTradeType(val);
-                setShareCount(1);
-              }
-            }}
-            fullWidth
-            color={tradeType === "buy" ? "success" : "error"}
-          >
-            <ToggleButton value="buy" disabled={user.id === profileUserId}>
-              Buy Shares
-            </ToggleButton>
-            <ToggleButton value="sell">Sell Shares</ToggleButton>
-          </ToggleButtonGroup>
-
-          <Box>
-            <Typography variant="subtitle2" sx={{ mb: 1 }}>
-              Number of Shares
-            </Typography>
-            <Stack direction="row" spacing={1} alignItems="center">
-              <Button
-                variant="outlined"
-                onClick={() => setShareCount((c) => Math.max(1, c - 1))}
-                sx={{ minWidth: 48 }}
-              >
-                -
-              </Button>
-              <TextField
-                fullWidth
-                type="number"
-                value={shareCount}
-                onChange={(e) => {
-                  const num = Number(e.target.value);
-                  if (num > 0) setShareCount(Math.floor(num));
-                }}
-                slotProps={{ htmlInput: { min: 1, style: { textAlign: "center" } } }}
-              />
-              <Button
-                variant="outlined"
-                onClick={() => setShareCount((c) => c + 1)}
-                sx={{ minWidth: 48 }}
-              >
-                +
-              </Button>
-            </Stack>
-            {tradeType === "sell" && (
-              <Typography variant="caption" color="text.secondary" display="block" sx={{ mt: 0.5 }}>
-                You own {stockInfo.sharesOwned} shares.
-              </Typography>
-            )}
-          </Box>
-
-          <Paper variant="outlined" sx={{ p: 2, background: "rgba(255,255,255,0.01)" }}>
-            <Stack spacing={1}>
-              <Stack direction="row" justifyContent="space-between">
-                <Typography variant="body2">Base Price:</Typography>
-                <Typography variant="body2" sx={{ fontWeight: "bold" }}>
-                  {tradePreview.price} Coins
-                </Typography>
-              </Stack>
-              <Stack direction="row" justifyContent="space-between">
-                <Typography variant="body2">Creator Fee (5%):</Typography>
-                <Typography variant="body2" color="success.main">
-                  +{tradePreview.creatorFee} Coins
-                </Typography>
-              </Stack>
-              <Stack direction="row" justifyContent="space-between">
-                <Typography variant="body2">System Fee (5%):</Typography>
-                <Typography variant="body2" color="warning.main">
-                  +{tradePreview.systemFee} Coins
-                </Typography>
-              </Stack>
-              <Box sx={{ borderTop: 1, borderColor: "divider", my: 1, pt: 1 }} />
-              <Stack direction="row" justifyContent="space-between">
-                <Typography sx={{ fontWeight: "bold" }}>
-                  {tradeType === "buy" ? "Total Cost:" : "Total Payout:"}
-                </Typography>
-                <Typography variant="h6" sx={{ fontWeight: "bold", color: "gold" }}>
-                  {tradePreview.total} Coins
-                </Typography>
-              </Stack>
-            </Stack>
-          </Paper>
-
-          {user.loggedIn && (
-            <Stack direction="row" justifyContent="space-between" sx={{ px: 1 }}>
-              <Typography variant="caption" color="text.secondary">
-                Coins After Trade:
-              </Typography>
-              <Typography
-                variant="caption"
-                sx={{
-                  fontWeight: "bold",
-                  color:
-                    tradeType === "buy"
-                      ? (user.coins || 0) >= tradePreview.total
-                        ? "success.main"
-                        : "error.main"
-                      : "success.main",
-                }}
-              >
-                {tradeType === "buy"
-                  ? ((user.coins || 0) - tradePreview.total).toFixed(2)
-                  : ((user.coins || 0) + tradePreview.total).toFixed(2)}{" "}
-                Coins
-              </Typography>
-            </Stack>
-          )}
-        </Stack>
-      </DialogContent>
-
-      <DialogActions sx={{ borderTop: 1, borderColor: "divider", p: 2 }}>
-        <Button onClick={() => setTradeModalOpen(false)} disabled={tradeSubmitting}>
-          Cancel
-        </Button>
-        <Button
-          variant="contained"
-          color={tradeType === "buy" ? "success" : "error"}
-          onClick={handleConfirmTrade}
-          disabled={
-            tradeSubmitting ||
-            shareCount <= 0 ||
-            (tradeType === "buy" && (user.coins || 0) < tradePreview.total) ||
-            (tradeType === "sell" && stockInfo.sharesOwned < shareCount)
-          }
-        >
-          {tradeSubmitting ? "Trading..." : tradeType === "buy" ? "Buy" : "Sell"}
-        </Button>
-      </DialogActions>
-    </Dialog>
-  );
 
   const buttonsBox = (
     <Grid
@@ -1725,7 +1465,19 @@ export default function Profile() {
 
   return (
     <>
-      {tradeModal}
+      <TradeDialog
+        open={tradeModalOpen}
+        onClose={() => setTradeModalOpen(false)}
+        stock={stockInfo ? {
+          targetType: "player",
+          id: profileUserId,
+          name: name,
+          avatar: avatar,
+          shareSupply: stockInfo.shareSupply,
+          sharesOwned: stockInfo.sharesOwned
+        } : null}
+        onSuccess={refetchProfile}
+      />
       {stats && (
         <StatsModal
           stats={stats}
@@ -1826,60 +1578,6 @@ export default function Profile() {
         </Grid>
         <Grid item xs={12} md={4}>
           <Stack direction="column" spacing={1}>
-            {stockInfo && (
-              <div className="box-panel" style={panelStyle}>
-                <div className="heading" style={headingStyle}>
-                  📈 Stock & Equity
-                </div>
-                <div className="content" style={{ display: "flex", flexDirection: "column", gap: "12px", padding: "16px" }}>
-                  <Stack direction="row" justifyContent="space-between" alignItems="center">
-                    <Box>
-                      <Typography variant="caption" color="text.secondary" display="block">
-                        Current Price
-                      </Typography>
-                      <Typography variant="h6" sx={{ fontWeight: "bold", color: "gold" }}>
-                        {stockInfo.buyPrice} Coins
-                      </Typography>
-                    </Box>
-                    <Box>
-                      <Sparkline history={stockInfo.priceHistory} />
-                    </Box>
-                  </Stack>
-
-                  <Grid container spacing={2} sx={{ mt: 0.5, borderTop: "1px solid rgba(255, 255, 255, 0.08)", pt: 1.5 }}>
-                    <Grid item xs={6}>
-                      <Typography variant="caption" color="text.secondary" display="block">
-                        Total Supply
-                      </Typography>
-                      <Typography variant="body2" sx={{ fontWeight: "bold" }}>
-                        {stockInfo.shareSupply} Shares
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={6}>
-                      <Typography variant="caption" color="text.secondary" display="block">
-                        Your Holdings
-                      </Typography>
-                      <Typography variant="body2" sx={{ fontWeight: "bold", color: stockInfo.sharesOwned > 0 ? "success.main" : "text.secondary" }}>
-                        {stockInfo.sharesOwned} Shares
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {!isSelf && user.loggedIn && (
-                    <Button
-                      variant="contained"
-                      color="warning"
-                      fullWidth
-                      onClick={handleOpenTrade}
-                      sx={{ mt: 1, fontWeight: "bold" }}
-                      startIcon={<i className="fas fa-chart-line" />}
-                    >
-                      Trade Shares
-                    </Button>
-                  )}
-                </div>
-              </div>
-            )}
             {mediaUrl && (
               <div className="box-panel" style={panelStyle}>
                 <MediaEmbed
@@ -1995,6 +1693,60 @@ export default function Profile() {
               <RapSheet userId={profileUserId} />
             )}
             {trophyCase}
+            {stockInfo && (
+              <div className="box-panel" style={panelStyle}>
+                <Typography variant="h3" sx={headingStyle}>
+                  Stock & Equity
+                </Typography>
+                <div className="content" style={{ display: "flex", flexDirection: "column", gap: "12px", padding: "16px" }}>
+                  <Stack direction="row" justifyContent="space-between" alignItems="center">
+                    <Box>
+                      <Typography variant="caption" color="text.secondary" display="block">
+                        Current Price
+                      </Typography>
+                      <Typography variant="h6" sx={{ fontWeight: "bold", color: "gold" }}>
+                        {stockInfo.buyPrice} Coins
+                      </Typography>
+                    </Box>
+                    <Box>
+                      <Sparkline history={stockInfo.priceHistory} />
+                    </Box>
+                  </Stack>
+
+                  <Grid container spacing={2} sx={{ mt: 0.5, borderTop: "1px solid rgba(255, 255, 255, 0.08)", pt: 1.5 }}>
+                    <Grid item xs={6}>
+                      <Typography variant="caption" color="text.secondary" display="block">
+                        Total Supply
+                      </Typography>
+                      <Typography variant="body2" sx={{ fontWeight: "bold" }}>
+                        {stockInfo.shareSupply} Shares
+                      </Typography>
+                    </Grid>
+                    <Grid item xs={6}>
+                      <Typography variant="caption" color="text.secondary" display="block">
+                        Your Holdings
+                      </Typography>
+                      <Typography variant="body2" sx={{ fontWeight: "bold", color: stockInfo.sharesOwned > 0 ? "success.main" : "text.secondary" }}>
+                        {stockInfo.sharesOwned} Shares
+                      </Typography>
+                    </Grid>
+                  </Grid>
+
+                  {!isSelf && user.loggedIn && (
+                    <Button
+                      variant="contained"
+                      color="warning"
+                      fullWidth
+                      onClick={handleOpenTrade}
+                      sx={{ mt: 1, fontWeight: "bold" }}
+                      startIcon={<i className="fas fa-chart-line" />}
+                    >
+                      Trade Shares
+                    </Button>
+                  )}
+                </div>
+              </div>
+            )}
             {hasAnyStats &&
               !settings.hideStatistics && (
                 <div className="box-panel ratings" style={panelStyle}>
